requirements.txt
torch==2.10.0
torchvision==0.25.0
pillow==12.1.1
matplotlib==3.10.8
lpips==0.1.4
faiss-cpu==1.13.2
tqdm==4.67.3
numpy==2.4.3
pyyaml==6.0.3